Fixed Pause / Start click calls.
Pause / Start button click events were bound to document instead of actual button due to on() selector containing selector and context.  This cause BOTH pause / start event functions being called at once. 

Not sure if we can use a context anymore or the proper way to use context in this sense ( when using document delegation ).

This might not be the best solution but it works now.

My first edit / pull on git so hope I am doing this properly :)